+++ examples/drawings/row-block.html
@@ -35,15 +35,6 @@
 
       openPlans.pencil.mode = "select";
 
-      const door = openPlans.door();
-      door.position.set(-2, 0.01, 0);
-
-      const window = openPlans.window();
-      window.position.set(2, 0.01, 0);
-
-      const wall = openPlans.wall();
-      wall.position.set(0, 0.01, 0);
-
       paperFrame = openPlans.paperFrame();
       paperFrame.format = 'A4';
       paperFrame.margin = 10;
